Deep update to solve minimist vulnerability
svg-to-png is dependent on the deprecated version on phantomjs-prebuilt, which hasn't been updated in 2 years. A couple weeks ago a vulnerability in minimist, which is a way down it's dependency tree, was published (with patches). Luckily many of these packages such as extract-zip and mkdirp that also hadn't update in a few years were quick to push updates for the new version (interestingly they also began active development again). For more information see GitHub advisory database: CVE-2020-7598